+++ src/components/PatronBlock/PatronBlockModalWithOverrideModal.test.js
@@ -27,16 +27,19 @@ describe('UserPermissions component', () => {
   it('Checking patronBlock Modal', () => {
     expect(renderPatronBlockModalWithOverrideModal(props)).toBeTruthy();
   });
+
   it('Checking patronBlock override Modal', () => {
     fireEvent.click(screen.getByText('ui-users.blocks.overrideButton'));
     fireEvent.click(document.querySelector('[data-test-override-patron-block-modal-close="true"]'));
     expect(screen.getByText('ui-users.blocks.modal.overridePatronBlock.header')).toBeTruthy();
   });
+
   it('Adding comment and saving', () => {
     fireEvent.click(screen.getByText('ui-users.blocks.overrideButton'));
-    fireEvent.change(document.querySelector('[id="textarea-input-2"]'), { target: {
-      value: 'test'
-    } });
+    fireEvent.change(
+      screen.getByLabelText('ui-users.blocks.modal.comment', { exact: false }),
+      { target: { value: 'test' } }
+    );
     fireEvent.click(document.querySelector('[data-test-override-patron-block-modal-save="true"]'));
     expect(screen.getByText('test')).toBeTruthy();
   });