+++ test/bigtest/interactors/KeyValue.js
@@ -1,9 +1,0 @@
-import {
-  interactor,
-  scoped,
-} from '@bigtest/interactor';
-
-export default interactor(class KeyValue {
-  label = scoped('div');
-  value = scoped('div:nth-child(2)');
-});