+++ test/bigtest/network/scenarios/fee-fine-actions.js
@@ -1,9 +1,0 @@
-export default (server) => {
-  server.get('/feefineactions', {
-    feefineactions: [
-      {
-        id: 'test-1',
-      },
-    ],
-  });
-};