+++ test/bigtest/network/scenarios/feefine-refund-error.js
@@ -1,7 +1,0 @@
-export default server => {
-  server.post('/feefine-reports/refund', {
-    errors: [{
-      message: 'error message',
-    }],
-  }, 422);
-};